Fix using the wrong remote nonce
It is possible that we use the wrong remote nonce in certain cases.

For example, in the following scenario, A and B are trying to commit and
acknowledge a normal channel update.

1: A sends a CommitmentSigned to B, which includes A's partial signature
   to spend the funding transaction output.
2: B checks the validity of CommitmentSigned. If valid, then sends a
   RevokeAndAck to A, which includes B's partial signature to spend the
   funding transaction output.
3: A checks the validity of the partial signature in the RevokeAndAck
   message.

In step 1, A needs a musig2 pubnonce sent from B to build a valid
partial signature for the CommitmentSigned message, call this pubnonce
previously_send_nonce_of_B (A also needs its own pubnonce, but that does
not matter too much to the problem we are illustrating). In step 2, B
also needs use this nonce to verify the validity of the partial signature.
In step 3, A still needs the same pubnonce to verify the vailidity of
the RevokeAndAck message sent by B. The problem is that for A,
this pubnonce is just the last nonce sent by B in a CommitmentSigned
message. This means if between step 1 and step 3, B sent a new
CommitmentSigned message to A, then the old previously_send_nonce_of_B
is overwritten. So while A is trying to verify the validity of the
RevokeAndAck message from B, he will find the partial signature invalid.

The solution to this problem is to save pub nonce used while
constructing the CommitmentSigned message. This is needed because the
node may receive serveral other CommitmentSigned message from the
counterparty, we wounld not be able to determine which pubnonce to use
if we have not saved the pubnonce in the first place.

A problem to this approach is that we need to save all the pubnonce sent
by B. This seems to be inevitable, as B may send as many commitment
signed as he likes to A while in step 1 and step 3. Of course, this is a
thing that no rational node will do.